+++ src/documentLinkProvider.ts
@@ -4,22 +4,26 @@ import * as vscode from 'vscode';
 import { ASMSymbolDocumenter } from './symbolDocumenter';
 
 export class ASMDocumentLinkProvider implements vscode.DocumentLinkProvider {
-  constructor(public symbolDocumenter: ASMSymbolDocumenter) {
+  constructor(public symbolDocumenter: ASMSymbolDocumenter) { }
 
-  }
-
-  provideDocumentLinks(document: vscode.TextDocument, token: vscode.CancellationToken): vscode.ProviderResult<vscode.DocumentLink[]> {
+  provideDocumentLinks(
+    document: vscode.TextDocument,
+    token: vscode.CancellationToken
+  ): vscode.ProviderResult<vscode.DocumentLink[]> {
     const filename = document.uri.fsPath;
     const table = this.symbolDocumenter.files[filename];
     if (table == null) {
       return undefined;
     }
 
-    let output: vscode.DocumentLink[] = [];
-    for (let i = 0; i < table.includedFiles.length; i++) {
-      const include = table.includedFiles[i];
-      if (include.fsPath) {
-        let link = new vscode.DocumentLink(include.range, vscode.Uri.file(include.fsPath));
+    const output: vscode.DocumentLink[] = [];
+    for (const include of table.includedFiles) {
+      const absolutePath = include.absolutePath;
+      if (absolutePath) {
+        const link = new vscode.DocumentLink(
+          include.range,
+          vscode.Uri.file(absolutePath)
+        );
         output.push(link);
       }
     }
@@ -27,7 +31,10 @@ export class ASMDocumentLinkProvider implements vscode.DocumentLinkProvider {
     return output;
   }
 
-  resolveDocumentLink(link: vscode.DocumentLink, token: vscode.CancellationToken): vscode.ProviderResult<vscode.DocumentLink> {
+  resolveDocumentLink(
+    link: vscode.DocumentLink,
+    token: vscode.CancellationToken
+  ): vscode.ProviderResult<vscode.DocumentLink> {
     return null;
   }
 }